requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
gunicorn==21.2.0